+++ engine/common/con_utils.c
@@ -566,7 +566,7 @@ qboolean Cmd_GetKeysList( const char *s, char *completedname, int length )
 		Con_Printf( "%16s\n", matchbuf );
 	}
 
-	Con_Printf( "\n^3 %lu keys found.\n", numkeys );
+	Con_Printf( "\n^3 %u keys found.\n", (uint32_t)numkeys );
 
 	if( completedname && length )
 	{
